refactor RepoGroup.vue into several smaller components
