+++ augurface/DevTodo.txt
@@ -1,3 +1,1 @@
-pull currently stored repositories and display
-
-implement repo groups (drag n drop, send to different group, etc)
+* Continue abstracting RepoGroup into several smaller components
